test: split monolithic controller tests into separate classes
