+++ src/test/java/roomescape/MissionStepTest.java
@@ -1,7 +1,6 @@
 package roomescape;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -11,7 +10,6 @@ import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -32,28 +30,6 @@ public class MissionStepTest {
     @Autowired
     private ReservationController reservationController;
 
-    @BeforeEach
-    void setUp() {
-        Map<String, String> time = new HashMap<>();
-        time.put("startAt", "10:00");
-
-        RestAssured.given().contentType(ContentType.JSON)
-                .body(time)
-                .when().post("/api/v1/times")
-                .then().statusCode(201);
-
-        Map<String, Object> themeParams = new HashMap<>();
-        themeParams.put("name", "이든의 공포 하우스");
-        themeParams.put("description", "이든이 귀신으로 나옴");
-        themeParams.put("imgUrl", "링크~");
-        themeParams.put("userName", "ADMIN");
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(themeParams)
-                .when().post("/api/v1/themes")
-                .then().statusCode(201);
-    }
-
     private Map<String, Object> reservationParams() {
         Map<String, Object> params = new HashMap<>();
         params.put("name", "브라운");
@@ -61,89 +37,6 @@ public class MissionStepTest {
         params.put("timeId", 1);
         params.put("themeId", 1);
         return params;
-    }
-
-    @Test
-    void 예약_조회() {
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-    }
-
-    @Test
-    void 예약_추가() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams())
-                .when().post("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(201)
-                .body("id", is(1));
-    }
-
-    @Test
-    void 예약_추가_조회() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams())
-                .when().post("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(201)
-                .body("id", is(1));
-
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("[0].id", is(1))
-                .body("[0].name", is("브라운"))
-                .body("[0].date", is("2023-08-05"))
-                .body("[0].time.id", is(1))
-                .body("[0].time.startAt", is("10:00"));
-    }
-
-    @Test
-    void 예약_추가_및_삭제() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams())
-                .when().post("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(201)
-                .body("id", is(1));
-
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-
-        RestAssured.given().log().all()
-                .when().delete("/api/v1/reservations/1")
-                .then().log().all()
-                .statusCode(204);
-
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-    }
-
-    @Test
-    void 존재하지_않는_예약을_삭제하면_404를_반환한다() {
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-
-        RestAssured.given().log().all()
-                .when().delete("/api/v1/reservations/1")
-                .then().log().all()
-                .statusCode(404);
     }
 
     @Test
@@ -159,6 +52,8 @@ public class MissionStepTest {
 
     @Test
     void DB_조회_API_전환() {
+        jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", "11:00");
+        jdbcTemplate.update("INSERT INTO theme (name, description, imgUrl) VALUES (?, ?, ?)", "놀이터", "테마놀이터", "image.html");
         jdbcTemplate.update("INSERT INTO reservation (name, date, time_id, theme_id) VALUES (?, ?, ?, ?)", "브라운",
                 "2023-08-05",
                 1, 1);
@@ -176,6 +71,9 @@ public class MissionStepTest {
 
     @Test
     void DB_추가_삭제_API_전환() {
+        jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", "11:00");
+        jdbcTemplate.update("INSERT INTO theme (name, description, imgUrl) VALUES (?, ?, ?)", "놀이터", "테마놀이터", "image.html");
+
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(reservationParams())
@@ -193,61 +91,6 @@ public class MissionStepTest {
 
         Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
         assertThat(countAfterDelete).isEqualTo(0);
-    }
-
-    @Test
-    void 시간_관리_API() {
-        RestAssured.given().log().all()
-                .when().get("/api/v1/times")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-
-        RestAssured.given().log().all()
-                .when().delete("/api/v1/times/1")
-                .then().log().all()
-                .statusCode(204);
-    }
-
-    @Test
-    void 예약과_시간_연결() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams())
-                .when().post("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(201);
-
-        RestAssured.given().log().all()
-                .when().get("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-    }
-
-    @Test
-    void 존재하지_않는_예약시간을_삭제하면_404를_반환한다() {
-        RestAssured.given().log().all()
-                .when().delete("/api/v1/times/2")
-                .then().log().all()
-                .statusCode(404);
-    }
-
-    @Test
-    void 예약이_존재하는_예약시간을_삭제하면_409를_반환한다() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservationParams())
-                .when().post("/api/v1/reservations")
-                .then().log().all()
-                .statusCode(201)
-                .body("id", is(1));
-
-        RestAssured.given().log().all()
-                .when().delete("/api/v1/times/1")
-                .then().log().all()
-                .statusCode(409);
-
     }
 
     @Test
